refactor: abstract qp_mark_outlier
The outlier marking and the subsequent usage of those marks to calculate
means are now two separate actions. `qp_calc_abs_mean` still functions
as normal in the case of no outlier marking prior (by calling the new
function) but if supplied with an `is_outlier` function, it uses
that (regardless if it's correct or not).